Organiza a pasta de Interface; Move a pasta de Database e corrige os erros originados por essas mudanças
