+++ src/Interface/Client/Model/login.fxml
@@ -8,9 +8,9 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<HBox prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.loginController">
+<HBox prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.Client.Controller.loginController">
    <children>
-      <VBox alignment="TOP_CENTER" prefWidth="250.0" spacing="5.0" styleClass="border-black" stylesheets="@style.css">
+      <VBox alignment="TOP_CENTER" prefWidth="250.0" spacing="5.0" styleClass="border-black" stylesheets="@../../style.css">
          <children>
             <Group>
                <children>
@@ -45,7 +45,7 @@
             <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
          </padding>
       </VBox>
-      <VBox alignment="TOP_CENTER" layoutX="14.0" prefWidth="550.0" spacing="5.0" styleClass="border-black" stylesheets="@style.css">
+      <VBox alignment="TOP_CENTER" layoutX="14.0" prefWidth="550.0" spacing="5.0" styleClass="border-black" stylesheets="@../../style.css">
          <children>
             <Group>
                <children>
