+++ src/Interface/Client/Model/menu.fxml
@@ -7,14 +7,14 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<VBox alignment="CENTER" prefHeight="600.0" prefWidth="800.0" spacing="20.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.menuController">
+<VBox alignment="CENTER" prefHeight="600.0" prefWidth="800.0" spacing="20.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.Client.Controller.menuController">
    <children>
       <Label fx:id="l_greetings" text="Olá, &lt;nome da pessoa&gt;">
          <font>
             <Font size="18.0" />
          </font>
       </Label>
-      <Button mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@style.css">
+      <Button mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@../../style.css">
          <font>
             <Font size="18.0" />
          </font>
@@ -44,7 +44,7 @@
             <Insets bottom="15.0" left="15.0" right="15.0" top="15.0" />
          </padding>
       </Button>
-      <Button layoutX="80.0" layoutY="257.0" mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@style.css">
+      <Button layoutX="80.0" layoutY="257.0" mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@../../style.css">
          <font>
             <Font size="18.0" />
          </font>
@@ -74,7 +74,7 @@
             <Insets bottom="15.0" left="15.0" right="15.0" top="15.0" />
          </padding>
       </Button>
-      <Button layoutX="540.0" layoutY="257.0" mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@style.css">
+      <Button layoutX="540.0" layoutY="257.0" mnemonicParsing="false" prefHeight="100.0" prefWidth="300.0" styleClass="default-btn" stylesheets="@../../style.css">
          <font>
             <Font size="18.0" />
          </font>
