+++ src/Interface/Server/Model/menu.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<VBox alignment="CENTER" prefHeight="600.0" prefWidth="800.0" spacing="20.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.menuController">
+<VBox alignment="CENTER" prefHeight="600.0" prefWidth="800.0" spacing="20.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.Client.Controller.menuController">
    <children>
       <Label prefHeight="27.0" text="Acesso como Servidor" textAlignment="CENTER">
          <font>
@@ -17,7 +17,7 @@
       </Label>
       <FlowPane alignment="CENTER" columnHalignment="CENTER" hgap="10.0" prefHeight="200.0" prefWidth="200.0" vgap="10.0">
          <children>
-            <Button mnemonicParsing="false" onAction="#listProducts" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@style.css">
+            <Button mnemonicParsing="false" onAction="#listProducts" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@../../style.css">
                <font>
                   <Font size="18.0" />
                </font>
@@ -47,7 +47,7 @@
                   <Insets bottom="15.0" left="15.0" right="15.0" top="15.0" />
                </padding>
             </Button>
-            <Button mnemonicParsing="false" onAction="#listClients" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@style.css">
+            <Button mnemonicParsing="false" onAction="#listClients" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@../../style.css">
                <font>
                   <Font size="18.0" />
                </font>
@@ -77,7 +77,7 @@
                   <Insets bottom="15.0" left="15.0" right="15.0" top="15.0" />
                </padding>
             </Button>
-            <Button mnemonicParsing="false" onAction="#listSell" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@style.css">
+            <Button mnemonicParsing="false" onAction="#listSell" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@../../style.css">
                <font>
                   <Font size="18.0" />
                </font>
@@ -107,7 +107,7 @@
                   <Insets bottom="15.0" left="15.0" right="15.0" top="15.0" />
                </padding>
             </Button>
-            <Button mnemonicParsing="false" onAction="#shutdownServer" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@style.css">
+            <Button mnemonicParsing="false" onAction="#shutdownServer" prefHeight="120.0" prefWidth="253.0" styleClass="default-btn" stylesheets="@../../style.css">
                <font>
                   <Font size="18.0" />
                </font>
