+++ src/Interface/Server/Model/productList.fxml
@@ -9,9 +9,9 @@
 <?import javafx.scene.layout.*?>
 <?fxml-stylesheet href="Interface/style.css"?>
 
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.productListController">
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Interface.Client.Controller.productListController">
    <children>
-      <HBox spacing="10.0" styleClass="title-bar" stylesheets="@style.css">
+      <HBox spacing="10.0" styleClass="title-bar" stylesheets="@../../style.css">
          <children>
             <Button mnemonicParsing="false" onAction="#backToMenu" styleClass="default-btn" text="Voltar">
                <graphic>
@@ -61,24 +61,24 @@
       </HBox>
       <HBox style="-fx-background-color: #444;">
          <children>
-            <Label alignment="CENTER" stylesheets="@style.css" text="ID">
+            <Label alignment="CENTER" stylesheets="@../../style.css" text="ID">
                <styleClass>
                   <String fx:value="table-header" />
                   <String fx:value="col-id" />
                </styleClass></Label>
-            <Label alignment="CENTER" layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="Produto" textAlignment="CENTER">
+            <Label alignment="CENTER" layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="Produto" textAlignment="CENTER">
                <styleClass>
                   <String fx:value="table-header" />
                   <String fx:value="col-product" />
                </styleClass>
             </Label>
-            <Label alignment="CENTER" layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="Preço" textAlignment="CENTER">
+            <Label alignment="CENTER" layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="Preço" textAlignment="CENTER">
                <styleClass>
                   <String fx:value="table-header" />
                   <String fx:value="col-price" />
                </styleClass>
             </Label>
-            <Label alignment="CENTER" layoutX="570.0" layoutY="10.0" stylesheets="@style.css" text="Qtd" textAlignment="CENTER">
+            <Label alignment="CENTER" layoutX="570.0" layoutY="10.0" stylesheets="@../../style.css" text="Qtd" textAlignment="CENTER">
                <styleClass>
                   <String fx:value="table-header" />
                   <String fx:value="col-amount" />
@@ -90,27 +90,27 @@
          <content>
             <VBox fx:id="vbox_list" prefHeight="200.0" prefWidth="100.0">
                <children>
-                  <HBox styleClass="table-line" stylesheets="@style.css">
+                  <HBox styleClass="table-line" stylesheets="@../../style.css">
                      <children>
-                        <Label alignment="CENTER" stylesheets="@style.css" text="12DA7">
+                        <Label alignment="CENTER" stylesheets="@../../style.css" text="12DA7">
                            <styleClass>
                               <String fx:value="col-id" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="Nome do Produto (fazer um FOR com essa linha; uma para cada produto)">
+                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="Nome do Produto (fazer um FOR com essa linha; uma para cada produto)">
                            <styleClass>
                               <String fx:value="col-product" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="10,99">
+                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="10,99">
                            <styleClass>
                               <String fx:value="col-price" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="610.0" layoutY="10.0" stylesheets="@style.css" text="7">
+                        <Label layoutX="610.0" layoutY="10.0" stylesheets="@../../style.css" text="7">
                            <styleClass>
                               <String fx:value="col-amount" />
                               <String fx:value="table-cell" />
@@ -118,27 +118,27 @@
                         </Label>
                      </children>
                   </HBox>
-                  <HBox layoutX="10.0" layoutY="10.0" styleClass="table-line" stylesheets="@style.css">
+                  <HBox layoutX="10.0" layoutY="10.0" styleClass="table-line" stylesheets="@../../style.css">
                      <children>
-                        <Label alignment="CENTER" stylesheets="@style.css" text="12DA7">
+                        <Label alignment="CENTER" stylesheets="@../../style.css" text="12DA7">
                            <styleClass>
                               <String fx:value="col-id" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="Nome do Produto (fazer um FOR com essa linha; uma para cada produto)">
+                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="Nome do Produto (fazer um FOR com essa linha; uma para cada produto)">
                            <styleClass>
                               <String fx:value="col-product" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@style.css" text="10,99">
+                        <Label layoutX="10.0" layoutY="10.0" stylesheets="@../../style.css" text="10,99">
                            <styleClass>
                               <String fx:value="col-price" />
                               <String fx:value="table-cell" />
                            </styleClass>
                         </Label>
-                        <Label layoutX="610.0" layoutY="10.0" stylesheets="@style.css" text="7">
+                        <Label layoutX="610.0" layoutY="10.0" stylesheets="@../../style.css" text="7">
                            <styleClass>
                               <String fx:value="col-amount" />
                               <String fx:value="table-cell" />
