+++ src/Server/Database/Sales.java
@@ -1,4 +1,4 @@
-package Database;
+package Server.Database;
 
 import Structure.Sale;
