+++ src/Server/Database/UsersDatabase.java
@@ -1,4 +1,4 @@
-package Database;
+package Server.Database;
 
 import Structure.User;
